refactor(services): relocate VC types to verifiable-credential module and remove legacy directories
