--- packages/services/src/adapters/index.ts
@@ -1,1 +1,0 @@
-export * from './verifiable-credentials';
