--- packages/services/src/adapters/verifiable-credentials/index.ts
@@ -1,1 +1,0 @@
-export * from './vckit.adapter.js';
