--- packages/services/src/interfaces/index.ts
@@ -1,2 +1,0 @@
-export * from './verifiableCredentialService.js';
-export type { StorageRecord } from './storageService.js';
